Move echo to commands package
